// SDKs/amazonaws.com/customer-profiles/2020-08-15/typescript/src/sdk/models/shared/putintegrationresponse.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class PutIntegrationResponse extends SpeakeasyBase {
  @Metadata({ data: "json, name=CreatedAt" })
  createdAt: Date;

  @Metadata({ data: "json, name=DomainName" })
  domainName: string;

  @Metadata({ data: "json, name=LastUpdatedAt" })
  lastUpdatedAt: Date;

  @Metadata({ data: "json, name=ObjectTypeName" })
  objectTypeName: string;

  @Metadata({ data: "json, name=Tags" })
  tags?: Map<string, string>;

  @Metadata({ data: "json, name=Uri" })
  uri: string;
}